requirements.txt
selenium
python-dotenv
schedule
pandas
openpyxl
sendgrid
